requirements.txt
openai==1.58.1
tqdm==4.67.1
tiktoken==0.12.0
lingua-language-detector==2.0.2
beautifulsoup4==4.12.0
html5lib==1.1
markitdown[all]==0.1.5
readability-lxml==0.8.4.1
python-dateutil==2.9.0
lxml==5.3.0
